# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Chat api documentation",
    "version": "1.0.0",
    "title": "Chat api"
  },
  "host": "https://chat-webka.herokuapp.com/api",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/rooms/{id}": {
      "delete": {
        "tags": [
          "Rooms"
        ],
        "summary": "Delete room by id",
        "operationId": "deleteRoom",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of room to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful"
          },
          "404": {
            "description": "Room not found"
          },
          "500": {
            "description": "Room not deleted, database error"
          }
        }
      },
      "patch": {
        "tags": [
          "Rooms"
        ],
        "summary": "Update room by ud",
        "operationId": "updateRoom",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of room to delete",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RoomUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful"
          },
          "404": {
            "description": "Room not found"
          },
          "500": {
            "description": "Room not updated, database error"
          }
        }
      },
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Get room's messages",
        "operationId": "getMessages",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of room to get messages",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Message"
              }
            }
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Create message in room",
        "operationId": "addMessage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of room to get messages",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MessageCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful"
          },
          "404": {
            "description": "Room not found"
          },
          "500": {
            "description": "Message not created, database error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Message": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "id of message"
        },
        "text": {
          "type": "string",
          "description": "content of message"
        }
      }
    },
...
    "RoomUpdate": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "name of room [unique]"
        }
      }
    },
    "MessageCreate": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string",
          "description": "content of message"
        }
      }
    },
...
  }
}